Refactor layout components: remove Header and Footer, add new Footer component, and update imports in Layout. Introduce DietSkeleton, IngredientSkeleton, and OpinionsSkeleton components for loading states. Remove unused RecipeDetails components and integrate IngredientResume for dietary information display.
