+++ src/components/layout/Footer.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Footer: React.FC = () => {
+    return (
+        <footer>
+            <p>© 2025 Projet GMF – Site réalisé dans le cadre d'une formation.</p>
+            <p><a href="">Mentions légales</a> | <a href="">Contact</a> | <a href="">Code source</a></p>
+            <p>Hébergé par <a href="" target="_blank"></a></p>
+        </footer>
+
+    );
+};
+
+export default Footer;
