+++ src/components/navBar/SearchBar.tsx
@@ -1,6 +1,6 @@
 import FilterAltIcon from '@mui/icons-material/FilterAlt';
 import SearchIcon from '@mui/icons-material/Search';
-import "../../../styles/nav.css";
+import "../../styles/nav.css";
 import { useNavigate } from 'react-router';
 
 interface SearchBarProps {
