+++ src/components/recipeDetails/DietSkeleton.tsx
