+++ src/components/recipeDetails/OpinionsResume.tsx
