+++ src/components/recipeDetails/StageResume.tsx
