--- src/dataFake/RecipeFake.json
@@ -1,132 +1,0 @@
-[
-  {
-    "id": 1,
-    "email": "user1@example.com",
-    "title": "Pâtes Carbonara",
-    "content": "Une recette italienne classique avec des œufs, du parmesan et du lard.",
-    "image": "https://gourmandiz.dhnet.be/wp-content/uploads/2018/03/carbonara-e1521461363147-690x388.jpg",
-    "person": 4,
-    "state": "published",
-    "rate": 4.5,
-    "nb_rate": 120,
-    "create": "2025-01-10",
-    "update": "2025-02-15"
-  },
-  {
-    "id": 2,
-    "email": "user2@example.com",
-    "title": "Poulet au curry miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam",
-    "content": "Un plat savoureux à base de poulet, lait de coco et épices. Miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam miam",
-    "image": "https://via.placeholder.com/150",
-    "person": 3,
-    "state": "draft",
-    "rate": 4.2,
-    "nb_rate": 85,
-    "create": "2025-02-01",
-    "update": "2025-02-10"
-  },
-  {
-    "id": 3,
-    "email": "user3@example.com",
-    "title": "Salade César",
-    "content": "Salade romaine, poulet grillé, croûtons et sauce César maison.",
-    "image": "https://via.placeholder.com/150",
-    "person": 1,
-    "state": "published",
-    "rate": 4.7,
-    "nb_rate": 200,
-    "create": "2025-01-20",
-    "update": "2025-02-18"
-  },
-  {
-    "id": 4,
-    "email": "user4@example.com",
-    "title": "Ratatouille",
-    "content": "Un mélange de légumes mijotés aux saveurs provençales.",
-    "image": "https://via.placeholder.com/150",
-    "person": 3,
-    "state": "published",
-    "rate": 4.8,
-    "nb_rate": 150,
-    "create": "2024-12-15",
-    "update": "2025-01-30"
-  },
-  {
-    "id": 5,
-    "email": "user5@example.com",
-    "title": "Bœuf bourguignon",
-    "content": "Un plat mijoté au vin rouge avec des légumes et du bœuf tendre.",
-    "image": "https://via.placeholder.com/150",
-    "person": 4,
-    "state": "published",
-    "rate": 4.6,
-    "nb_rate": 95,
-    "create": "2024-11-10",
-    "update": "2025-01-12"
-  },
-  {
-    "id": 6,
-    "email": "user6@example.com",
-    "title": "Quiche Lorraine",
-    "content": "Une quiche savoureuse aux lardons et fromage.",
-    "image": "https://via.placeholder.com/150",
-    "person": 4,
-    "state": "draft",
-    "rate": 4.3,
-    "nb_rate": 80,
-    "create": "2025-01-25",
-    "update": "2025-02-14"
-  },
-  {
-    "id": 7,
-    "email": "user7@example.com",
-    "title": "Tarte aux pommes",
-    "content": "Une délicieuse tarte avec des pommes caramélisées.",
-    "image": "https://via.placeholder.com/150",
-    "person": 3,
-    "state": "published",
-    "rate": 4.9,
-    "nb_rate": 210,
-    "create": "2024-12-05",
-    "update": "2025-02-01"
-  },
-  {
-    "id": 8,
-    "email": "user8@example.com",
-    "title": "Chili con carne",
-    "content": "Un plat épicé à base de viande hachée, haricots rouges et tomates.",
-    "image": "https://via.placeholder.com/150",
-    "person": 4,
-    "state": "published",
-    "rate": 4.4,
-    "nb_rate": 130,
-    "create": "2025-01-08",
-    "update": "2025-02-09"
-  },
-  {
-    "id": 9,
-    "email": "user9@example.com",
-    "title": "Soupe à l'oignon",
-    "content": "Une soupe traditionnelle gratinée avec du fromage.",
-    "image": "https://via.placeholder.com/150",
-    "person": 3,
-    "state": "published",
-    "rate": 4.5,
-    "nb_rate": 90,
-    "create": "2025-01-15",
-    "update": "2025-02-11"
-  },
-  {
-    "id": 10,
-    "email": "user10@example.com",
-    "title": "Mousse au chocolat",
-    "content": "Un dessert aérien et gourmand au chocolat noir qui ravira les petits et les grands.",
-    "image": "https://via.placeholder.com/150",
-    "person": 2,
-    "state": "published",
-    "rate": 4.9,
-    "nb_rate": 250,
-    "create": "2025-01-02",
-    "update": "2025-02-20"
-  }
-]
