--- src/dataFake/dataCategory.json
@@ -1,15 +1,0 @@
-[
-    "Produits laitiers",
-    "Viandes",
-    "Poissons",
-    "Légumes",
-    "Fruits",
-    "Céréales",
-    "Féculents",
-    "Sucreries",
-    "Boissons",
-    "Épices",
-    "Condiments",
-    "Sauces",
-    "Autres"
-]
