--- src/pages/1_header/Header.tsx
@@ -1,11 +1,0 @@
-import {FC} from 'react';
-
-const Header: FC<{}> = ({}) => {
-    return (
-        <>
-
-        </>
-    );
-};
-
-export default Header;
