--- src/pages/3_footer/Footer.tsx
@@ -1,12 +1,0 @@
-import {FC} from 'react';
-import "../../styles/footer.css";
-
-const Footer: FC<{}> = ({}) => {
-    return (
-        <div className="footer">
-                Footer du site
-        </div>
-    );
-};
-
-export default Footer;
